Migrate AggProfileRepo and related models to root g2 package
Moved `ProfileData`, `AggProfileRepo`, and `AggProfile` from `cmd/g2/site.go` to the root `site.go` in the `g2` package.
Updated all dependent structs and usages in `cmd/g2/*` to point to the new `g2.*` types.
Fixed imports and ensured all tests/lints pass successfully.
